move updating ID in getASelfing and misc
(1) In getASelfing() we had a bug, when we create the new ID, paste(id,cycles,sep=sepChar), we need to updates those IDs in the sire and dam columns. We were doing those updates only for the individuals that had nCycles>1. We need to do this update always regardless of the number of cycles of selfing.

(2) Change the sepChar default to be -S standing for 'Selfing'

(3) Change the condition if(ncycles>1) to be if(ncycles>0). The interpretation is that nCycles is the nubmer of selfing cycles.